ai_services/requirements.txt
openai==1.12.0
chromadb==0.4.22
sentence-transformers==2.3.1
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
python-dotenv==1.0.0
numpy==1.26.3
